08/29 19:42

+++ JamUtils/ShardSlot.h
@@ -11,7 +11,7 @@ namespace jam::utils::exec
     	COUNT
     };
 
-    enum class eShardState : uint32
+    enum class eShardState : uint8
     {
 	    CLOSED      = 0,
     	OPEN        = 1,
@@ -23,7 +23,7 @@ namespace jam::utils::exec
 	{
         std::atomic<Mailbox*>   q{ nullptr };                       // 현재 게시된 큐(없으면 nullptr)
         std::atomic<uint32>     gen{ 0 };                              // 세대 카운터(ABA/교체 감지)
-        std::atomic<uint32>     state{ E2U(eShardState::CLOSED) };         // 운영 상태
+        std::atomic<uint8>      state{ E2U(eShardState::CLOSED) };         // 운영 상태
     };
 
     struct ShardSlot
